requirements.txt
streamlit
google-genai
python-dotenv
Flask
twilio
asyncpg
langdetect
sendgrid